# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Rivus API
  description: 多媒体管理系统的后端 API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 本地开发环境
paths:
  /ping:
    get:
      operationId: Ping
      summary: 健康检查接口
      description: 用于测试服务是否正常运行
      responses:
        '200':
          description: 服务正常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PingResponse'
  /config:
    get:
      operationId: GetConfig
      summary: 获取配置接口
      description: 用于获取服务配置
      responses:
        '200':
          description: 配置成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetConfigResponse'
    post:
      operationId: SetConfig
      summary: 配置接口
      description: 用于配置服务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetConfigRequest'
      responses:
        '200':
          description: 配置成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetConfigResponse'
  /browse:
    get:
      operationId: Browse
      summary: 浏览本地目录
      description: 用于浏览本地目录
      parameters:
        - name: directory
          in: query
          schema:
            type: string
            default: /
            description: 本地目录路径，默认根目录
      responses:
        '200':
          description: 浏览成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrowseResponse'
  /audio/scan:
    get:
      operationId: ScanAudio
      summary: 扫描音频目录，构建音频元数据
      description: 根据文件存储目录初始化服务数据
      responses:
        '200':
          description: 扫描成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanAudioResponse'
  /audio/search:
    get:
      operationId: SearchAudio
      summary: 搜索音频
      description: 根据音频元数据搜索音频
      parameters:
        # 分页参数
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            description: 页码，默认第一页
          required: false
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
            description: 每页数量，默认10条
          required: false
        # 排序参数
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [title, artist]
            default: title
            description: 排序字段，默认按标题排序
          required: false
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
            description: 排序顺序，默认升序
          required: false
        # 搜索参数
        - name: title
          in: query
          schema:
            type: string
            default: ''
            description: 搜索查询，默认空字符串
          required: false
        - name: artist
          in: query
          schema:
            type: string
            default: ''
            description: 搜索查询，默认空字符串
          required: false
      responses:
        '200':
          description: 搜索成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchAudioResponse'
  /audio/delete:
    delete:
      operationId: DeleteAudio
      summary: 删除音频
      description: 根据音频ID删除音频
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteAudioRequest'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteAudioResponse'
  /audio/restore:
    post:
      operationId: RestoreAudio
      summary: 恢复音频
      description: 根据音频ID恢复音频
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestoreAudioRequest'
      responses:
        '200':
          description: 恢复成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestoreAudioResponse'
  /audio/update:
    post:
      operationId: UpdateAudio
      summary: 更新音频
      description: 根据音频ID更新音频元数据
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAudioRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateAudioResponse'
  /audio/recyclebin:
    get:
      operationId: GetRecycleBin
      summary: 获取回收站音频
      description: 获取所有已删除的音频
      parameters:
        # 分页参数
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            description: 页码，默认第一页
          required: false
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
            description: 每页数量，默认10条
          required: false
        # 排序参数
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [title, artist]
            default: title
            description: 排序字段，默认按标题排序
          required: false
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
            description: 排序顺序，默认升序
          required: false
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetRecycleBinResponse'
  /audio/source:
    get:
      x-codegen-skip: true
      operationId: GetAudioSource
      summary: 获取音频流
      description: 获取音频流
      parameters:
        - name: id
          in: query
          schema:
            type: integer
            default: 1
            description: 音频ID
          required: true
      responses:
        '200':
          description: 音频分片
          content:
            audio/mpeg: {}
            audio/flac: {}
  /audio/upload:
    post:
      x-codegen-skip: true
      operationId: UploadAudio
      summary: 上传音频文件
      description: 上传音频文件到服务器
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: 音频文件
      responses:
        '200':
          description: 上传成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
  /audio/lyric/upload:
    post:
      x-codegen-skip: true
      operationId: UploadLyric
      summary: 上传歌词文件
      description: 上传歌词文件到服务器
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadLyricsRequest'
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: 歌词文件
      responses:
        '200':
          description: 上传成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
  /audio/scrape:
    get:
      operationId: ScrapeAudio
      summary: 刮削音频元数据
      description: 通过 AcoustID 和 MusicBrainz 刮削音频元数据（标题、艺术家、专辑）
      parameters:
        - name: id
          in: query
          schema:
            type: integer
            default: 1
            description: 音频ID
          required: true
      responses:
        '200':
          description: 刮削成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScrapeAudioResponse'
  /audio/lyric/fetch:
    get:
      operationId: FetchLyric
      summary: 拉取歌词
      description: 通过 LRCLIB 根据音频元数据拉取时间同步歌词（LRC格式）
      parameters:
        - name: id
          in: query
          schema:
            type: integer
            default: 1
            description: 音频ID
          required: true
      responses:
        '200':
          description: 拉取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FetchLyricResponse'
  /audio/lyric/source:
    get:
      x-codegen-skip: true
      operationId: GetLyricSource
      summary: 获取歌词
      description: 获取音频的歌词
      parameters:
        - name: id
          in: query
          schema:
            type: integer
            default: 1
            description: 音频ID
          required: true
      responses:
        '200':
          description: 获取成功
          content:
            text/plain: {}
  /audio/sync/test-connection:
    post:
      operationId: SyncTestConnection
      description: 测试FTP连接与凭据是否有效
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncTestConnectionRequest'
      responses:
        '200':
          description: 测试成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncTestConnectionResponse'
  /audio/sync/browse:
    get:
      operationId: SyncBrowse
      description: 浏览FTP服务器目录
      parameters:
        - name: directory
          in: query
          schema:
            type: string
            default: /
            description: FTP目录路径
        - name: ip
          in: query
          schema:
            type: string
            default: 192.168.1.100
            description: FTP服务器IP地址
        - name: port
          in: query
          schema:
            type: integer
            default: 21
            description: FTP服务器端口
        - name: username
          in: query
          schema:
            type: string
            default: admin
            description: FTP用户名
        - name: password
          in: query
          schema:
            type: string
            default: 123456
            description: FTP密码
      responses:
        '200':
          description: 浏览成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncBrowseResponse'
  /audio/sync/compare:
    post:
      operationId: SyncCompare
      description: 对比本地音频文件与FTP服务器音频文件
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncCompareRequest'
      responses:
        '200':
          description: 对比成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncCompareResponse'
  /audio/sync/execute:
    post:
      operationId: SyncExecute
      description: 执行同步操作
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncExecuteRequest'
      responses:
        '200':
          description: 执行成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncExecuteResponse'
    

components:
  schemas:
    PingResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
    Config:
      type: object
      properties:
        audio_dir:
          type: string
          example: /c/users/default/Music
          description: 音频目录
        acoustid_api_key:
          type: string
          description: AcoustID API Key
    SetConfigRequest:
      type: object
      $ref: '#/components/schemas/Config'
    SetConfigResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
    GetConfigResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          type: object
          $ref: '#/components/schemas/Config'
    BrowseResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          type: object
          properties:
            separator:
              type: string
              description:  分隔符
            dirs:
              type: array
              items:
                type: string
              description: 本地目录列表
    ScanAudioResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
    AudioItem:
      type: object
      properties:
        id:
          type: integer
          description: ID
        create_time:
          type: string
          description: 创建时间
        update_time:
          type: string
          description: 更新时间
        delete_time:
          type: string
          description: 删除时间
        file_path:
          type: string
          description: 音频文件路径
        file_ext:
          type: string
          description: 音频文件扩展名
        file_size:
          type: integer
          description: 音频文件大小，单位字节
        file_md5:
          type: string
          description: 音频文件MD5值
        title:
          type: string
          description: 标题
        artist:
          type: string
          description: 艺术家
        album:
          type: string
          description: 专辑
        duration:
          type: number
          description: 时长（秒）
        cover_path:
          type: string
          description: 封面路径
        lyric_path:
          type: string
          description: 歌词路径
    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
          description: 总记录数
        total_pages:
          type: integer
          description: 总页数
    SearchAudioResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          type: object
          properties:
            audio_list:
              type: array
              items:
                $ref: '#/components/schemas/AudioItem'
            pagination:
              $ref: '#/components/schemas/Pagination'
    DeleteAudioRequest:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
          description: 音频ID列表
        hard:
          type: boolean
          description: 是否硬删除，默认false
          default: false
    DeleteAudioResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
    RestoreAudioRequest:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
          description: 音频ID列表
    RestoreAudioResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
    UpdateAudioRequest:
      type: object
      properties:
        id:
          type: integer
          description: 音频ID
        title:
          type: string
          description: 标题
        artist:
          type: string
          description: 艺术家
        album:
          type: string
          description: 专辑
    UpdateAudioResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
    GetRecycleBinResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          type: object
          properties:
            audio_list:
              type: array
              items:
                $ref: '#/components/schemas/AudioItem'
            pagination:
              $ref: '#/components/schemas/Pagination'
    UploadLyricsRequest:
      type: object
      properties:
        id:
          type: integer
          description: 音频ID
    UploadResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
    ScrapeAudioResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          type: object
          properties:
            title:
              type: string
              description: 歌曲标题
            artist:
              type: string
              description: 艺术家
            album:
              type: string
              description: 专辑
    FetchLyricResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          type: object
          properties:
            synced_lyrics:
              type: string
              description: 时间同步歌词（LRC格式，已转为简体中文）
    SyncTestConnectionRequest:
      type: object
      properties:
        ip:
          type: string
          description: FTP IP地址
        port:
          type: integer
          description: FTP端口号
        username:
          type: string
          description: FTP用户名
        password:
          type: string
          description: FTP密码
    SyncTestConnectionResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
    SyncBrowseResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          type: object
          properties:
            dirs:
              type: array
              items:
                type: string
              description: FTP服务器目录列表
    SyncCompareRequest:
      type: object
      properties:
        ip:
          type: string
          description: FTP IP地址
        port:
          type: integer
          description: FTP端口号
        username:
          type: string
          description: FTP用户名
        password:
          type: string
          description: FTP密码
        directory:
          type: string
          description: FTP目录路径
    SyncCompareResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          type: object
          properties:
            to_download:
              type: array
              items:
                type: string
              description: 待下载的音频文件列表
            to_upload:
              type: array
              items:
                type: string
              description: 待上传的音频文件列表
            conflict:
              type: array
              items:
                type: string
              description: 冲突的音频文件列表
            unchanged:
              type: integer
              description: 未改变的音频文件数量
    SyncExecuteRequest:
      type: object
      properties:
        ip:
          type: string
          description: FTP IP地址
        port:
          type: integer
          description: FTP端口号
        username:
          type: string
          description: FTP用户名
        password:
          type: string
          description: FTP密码
        directory:
          type: string
          description: FTP目录路径
        to_download:
          type: array
          items:
            type: string
          description: 待下载的音频文件列表
        to_upload:
          type: array
          items:
            type: string
          description: 待上传的音频文件列表
    SyncExecuteResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success