# internal/handler/deleteAudio.handler.go
// Code generated by GinX. DO NOT EDIT.
package handler

import (
	"net/http"

	generated "Rivus/internal/generated"
	"Rivus/internal/logic"

	"github.com/gin-gonic/gin"
)

// DeleteAudio 删除音频
func (h *Handler) DeleteAudio(c *gin.Context) {
	l := logic.NewDeleteAudioLogic(h.Logic)

	var req generated.DeleteAudioRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	resp, err := l.DeleteAudio(&req)

	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, resp)
}