# internal/handler/setConfig.handler.go
// Code generated by GinX. DO NOT EDIT.
package handler

import (
	"net/http"

	generated "Rivus/internal/generated"
	"Rivus/internal/logic"

	"github.com/gin-gonic/gin"
)

// SetConfig 配置接口
func (h *Handler) SetConfig(c *gin.Context) {
	l := logic.NewSetConfigLogic(h.Logic)

	var req generated.SetConfigRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	resp, err := l.SetConfig(&req)

	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, resp)
}