# internal/handler/syncCompare.handler.go
// Code generated by GinX. DO NOT EDIT.
package handler

import (
	"net/http"

	generated "Rivus/internal/generated"
	"Rivus/internal/logic"

	"github.com/gin-gonic/gin"
)

// SyncCompare
func (h *Handler) SyncCompare(c *gin.Context) {
	l := logic.NewSyncCompareLogic(h.Logic)

	var req generated.SyncCompareRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	resp, err := l.SyncCompare(&req)

	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, resp)
}