# internal/handler/syncExecute.handler.go
// Code generated by GinX. DO NOT EDIT.
package handler

import (
	"net/http"

	generated "Rivus/internal/generated"
	"Rivus/internal/logic"

	"github.com/gin-gonic/gin"
)

// SyncExecute
func (h *Handler) SyncExecute(c *gin.Context) {
	l := logic.NewSyncExecuteLogic(h.Logic)

	var req generated.SyncExecuteRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	resp, err := l.SyncExecute(&req)

	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, resp)
}