# internal/handler/syncTestConnection.handler.go
// Code generated by GinX. DO NOT EDIT.
package handler

import (
	"net/http"

	generated "Rivus/internal/generated"
	"Rivus/internal/logic"

	"github.com/gin-gonic/gin"
)

// SyncTestConnection
func (h *Handler) SyncTestConnection(c *gin.Context) {
	l := logic.NewSyncTestConnectionLogic(h.Logic)

	var req generated.SyncTestConnectionRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	resp, err := l.SyncTestConnection(&req)

	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, resp)
}